requirements.txt
numpy
scipy
pandas
scikit-learn
matplotlib
tqdm
rich
pytorch-lightning==1.8.6
hydra-core
omegaconf
wandb
einops
opt_einsum
cmake # For pykeops support
# pykeops # Only for S4D. If there are installation problems with pykeops==2.x, try pykeops==1.5
transformers==4.26.1 # For some schedulers and tokenizers
torchvision
timm
prettytable
numerize
git-lfs

# Dataset specific packages
torchtext # this needs to align with the pytorch version
datasets # LRA

# genomic specific
pyfaidx
polars
genomic-benchmarks
loguru
liftover
